git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@12344 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/cockpit.h
@@ -243,8 +243,8 @@ class CGenericCockpit {
 		inline CCockpitInfo& Info (void) { return m_info; }
 		inline int Type (void) { return m_info.nType; }
 		inline int LineSpacing (void) { return m_info.nLineSpacing = GAME_FONT->Height () + GAME_FONT->Height () / 4; }
-		inline float XScale (void) { return m_info.xScale = screen.Scale (0); }
-		inline float YScale (void) { return m_info.yScale = screen.Scale (1); }
+		inline float XScale (void) { return m_info.xScale = (float) screen.Scale (0); }
+		inline float YScale (void) { return m_info.yScale = (float) screen.Scale (1); }
 		inline void SetScales (float xScale, float yScale) { m_info.xScale = xScale, m_info.yScale = yScale; }
 		inline void Rebuild (void) { m_info.bRebuild = true; }
